refactor(oauth): extract storage lifecycle
